+++ src/main/java/faang/school/projectservice/config/s3/S3Config.java
@@ -3,6 +3,7 @@ package faang.school.projectservice.config.s3;
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
@@ -13,20 +14,19 @@ import org.springframework.context.annotation.Configuration;
 public class S3Config {
     @Value("${services.s3.accessKey}")
     private String accessKey;
-
     @Value("${services.s3.secretKey}")
     private String secretKey;
-
     @Value("${services.s3.region}")
     private String region;
-
+    @Value("${services.s3.endpoint}")
+    private String endpoint;
 
     @Bean
     public AmazonS3 S3Client() {
-        AWSCredentials credentials = new BasicAWSCredentials(accessKey, secretKey);
         return AmazonS3ClientBuilder.standard()
-                .withCredentials(new AWSStaticCredentialsProvider(credentials))
-                .withRegion(region)
+                .withEndpointConfiguration(new AwsClientBuilder.EndpointConfiguration(endpoint, region))
+                .withCredentials(new AWSStaticCredentialsProvider(new BasicAWSCredentials(accessKey, secretKey)))
+                .withPathStyleAccessEnabled(true)
                 .build();
     }
 }
