+++ src/main/java/faang/school/projectservice/service/s3/S3Service.java
@@ -3,23 +3,37 @@ package faang.school.projectservice.service.s3;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import faang.school.projectservice.model.Project;
+import faang.school.projectservice.repository.ProjectRepository;
+import faang.school.projectservice.validator.project.ProjectValidator;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @Slf4j
+@RequiredArgsConstructor
 public class S3Service {
     private final AmazonS3 s3client;
-    @Autowired
-    public S3Service(AmazonS3 s3client) {
-        this.s3client = s3client;
-    }
+    private final ProjectRepository projectRepository;
+    private final ProjectValidator validator;
 
     @Value("${services.s3.bucketName}")
     private String bucketName;
+
+    public String addCover(Long projectId, MultipartFile coverImage) {
+        Project project = validator.getProjectAfterValidateId(projectId);
+        validator.validateOwnerId(project);
+
+        String folder = project.getId() + project.getName();
+        String key = uploadFile(coverImage, folder);
+        project.setCoverImageId(key);
+        projectRepository.save(project);
+
+        return "Cover Image added to project " + projectId;
+    }
 
     public String uploadFile(MultipartFile file, String folder) {
         ObjectMetadata objectMetadata = new ObjectMetadata();
